verifier: Support running without a wallet
The intended use case is that a user can create an SPV with wallet=None,
and call _request_and_verify_single_proof, which will raise an exception
if the verification failed.  In this use case, no data about the verified
transactions will be saved to a wallet.  The user is responsible for
making sure that the relevant headers have already been downloaded via
request_chunk.

From Bitcoin PR 5660