Fix handling of secondary connections that were never made
Connections that were never made weren't added to interfaces_clean, so we
were incorrectly interpreting them as dirty and therefore failing to add
them to the disconnected list.  This caused the connection to be retried
immediately instead of waiting a while, which put an unreasonable load on
the Tor network.  We now instead check whether they are in the dirty list,
which should work fine.